feat: logs pause
